[refactor] ArticleNode, FolderNode를 바깥으로 이동 #4 - Reaction 엔티티의 이름을 Emoji로 변경
